[NodeTypeResolver] decouple nested call types Test
