+++ packages/NodeTypeResolver/tests/NodeCallerTypeResolver/NestedMethodCallSource/NestedMethodCalls.php.inc
