+++ packages/NodeTypeResolver/tests/NodeCallerTypeResolver/NestedMethodCallSource/OnMethodCallCallDifferentType.php.inc
