+++ src/Rector/Contrib/Nette/Forms/FormSetRequiredRector.php
@@ -38,7 +38,11 @@ final class FormSetRequiredRector extends AbstractRector
 
     public function isCandidate(Node $node): bool
     {
-        if (! $this->methodCallAnalyzer->isTypeAndMethods($node, self::FORM_CLASS, ['addCondition'])) {
+        if (! $this->methodCallAnalyzer->isTypesAndMethods(
+            $node,
+            [self::FORM_CLASS, 'Nette\Forms\Container'],
+            ['addCondition'])
+        ) {
             return false;
         }
